saas_server/requirements.txt
fastapi>=0.115.0
uvicorn>=0.32.0
python-multipart>=0.0.12
python-dotenv>=1.0.0
supabase>=2.10.0
pydantic>=2.10.0
aiofiles>=24.1.0
